+++ src/api/v1/profile/profile.constant.ts
@@ -2,4 +2,5 @@ export const success = {
   PROFILE_UPDATED_SUCCESSFULLY: "Profile updated successfully",
   PROFILE_FETCHED_SUCCESSFULLY: "Profile fetched successfully",
   PROFILE_VISIBILITY_UPDATED_SUCCESSFULLY: "Profile visibility updated successfully",
+  PROFILE_PERMISSIONS_FETCHED_SUCCESSFULLY: "Profile permissions fetched successfully",
 };