+++ src/utils/test/profile.utils.ts
@@ -8,7 +8,7 @@ import { app } from "../../app.js";
 
 export const upSertProfileData = async (data: CreateProfileData, authorizationHeader: string) => {
   return await supertest(app)
-    .put(`${ROUTES.API_ROOT}${ROUTES.PROFILES.ROOT}`)
+    .put(`${ROUTES.API_ROOT}${ROUTES.PROFILE.ROOT}`)
     .set("authorization", authorizationHeader)
     .send(data);
 };
@@ -37,7 +37,7 @@ export const expectProfileData = (response: Response, data: ProfileData) => {
 
 export const toggleProfileVisibility = async (isPublic: boolean, authorizationHeader: string) => {
   return await supertest(app)
-    .patch(`${ROUTES.API_ROOT}${ROUTES.PROFILES.VISIBILITY}`)
+    .patch(`${ROUTES.API_ROOT}${ROUTES.PROFILE.VISIBILITY}`)
     .set("authorization", authorizationHeader)
     .send({ isPublic });
 };
@@ -49,4 +49,17 @@ export const expectProfileVisibility = (response: Response, isPublic: boolean) =
   expect(response.body.status).toBe("success");
 
   expect(response.body.data.isPublic).toBe(isPublic);
+};
+
+export const getProfileData = async (authorizationHeader: string) => {
+  return await supertest(app).get(`${ROUTES.API_ROOT}${ROUTES.PROFILE.ROOT}`).set("authorization", authorizationHeader);
+};
+
+export const expectGetProfileData = (response: Response, data: ProfileData) => {
+  expect(response).toBeDefined();
+  expect(response.statusCode).toBe(STATUS_CODES.OK);
+  expect(response.body.message).toBe(success.PROFILE_FETCHED_SUCCESSFULLY);
+  expect(response.body.status).toBe("success");
+
+  expect(response.body.data).toMatchObject(data);
 };